[Refactoring] Move Authenticator from mailbox-store to mailbox-api
